backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
django-filter==2.3.0
Pillow==10.4.0
PyYAML==6.0
gunicorn==20.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
drf-extra-fields==3.7.0

# Дополнительные зависимости для улучшений
redis==4.5.4
django-redis==5.2.0
drf-yasg==1.21.5
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
